# Tune Arithmetic Op launch specification