--- a/dali/operators/math/expressions/expression_impl_gpu.cuh
+++ b/dali/operators/math/expressions/expression_impl_gpu.cuh
@@ -33,12 +33,15 @@ namespace dali {
 template <ArithmeticOp op, typename Result, typename Input>
 __device__ void ExecuteUnOp(Result *result, const Input *in, int64_t extent) {
   using meta = arithm_meta<op, GPUBackend>;
-  result += blockIdx.x * blockDim.x + threadIdx.x;
-  in += blockIdx.x * blockDim.x + threadIdx.x;
-  for (int64_t i = blockIdx.x * blockDim.x + threadIdx.x; i < extent; i += blockDim.x * gridDim.x) {
+  int64_t start_ofs = static_cast<int64_t>(blockDim.x) * blockIdx.x + threadIdx.x;
+  int64_t stride = static_cast<int64_t>(blockDim.x) * gridDim.x;
+  auto *tile_end = result + extent;
+  result += start_ofs;
+  in += start_ofs;
+  while (result < tile_end) {
     *result = meta::impl(*in);
-    result += blockDim.x * gridDim.x;
-    in += blockDim.x * gridDim.x;
+    result += stride;
+    in += stride;
   }
 }
 
@@ -48,14 +51,17 @@ __device__ void ExecuteUnOp(Result *result, const Input *in, int64_t extent) {
 template <ArithmeticOp op, typename Result, typename Left, typename Right>
 __device__ void ExecuteBinOp(Result *result, const Left *l, const Right *r, int64_t extent) {
   using meta = arithm_meta<op, GPUBackend>;
-  result += blockIdx.x * blockDim.x + threadIdx.x;
-  l += blockIdx.x * blockDim.x + threadIdx.x;
-  r += blockIdx.x * blockDim.x + threadIdx.x;
-  for (int64_t i = blockIdx.x * blockDim.x + threadIdx.x; i < extent; i += blockDim.x * gridDim.x) {
+  int64_t start_ofs = static_cast<int64_t>(blockDim.x) * blockIdx.x + threadIdx.x;
+  int64_t stride = static_cast<int64_t>(blockDim.x) * gridDim.x;
+  auto *tile_end = result + extent;
+  result += start_ofs;
+  l += start_ofs;
+  r += start_ofs;
+  while (result < tile_end) {
     *result = meta::impl(*l, *r);
-    result += blockDim.x * gridDim.x;
-    l += blockDim.x * gridDim.x;
-    r += blockDim.x * gridDim.x;
+    result += stride;
+    l += stride;
+    r += stride;
   }
 }
 
@@ -65,12 +71,15 @@ __device__ void ExecuteBinOp(Result *result, const Left *l, const Right *r, int6
 template <ArithmeticOp op, typename Result, typename Left, typename Right>
 __device__ void ExecuteBinOp(Result *result, Left l, const Right *r, int64_t extent) {
   using meta = arithm_meta<op, GPUBackend>;
-  result += blockIdx.x * blockDim.x + threadIdx.x;
-  r += blockIdx.x * blockDim.x + threadIdx.x;
-  for (int64_t i = blockIdx.x * blockDim.x + threadIdx.x; i < extent; i += blockDim.x * gridDim.x) {
+  int64_t start_ofs = static_cast<int64_t>(blockDim.x) * blockIdx.x + threadIdx.x;
+  int64_t stride = static_cast<int64_t>(blockDim.x) * gridDim.x;
+  auto *tile_end = result + extent;
+  result += start_ofs;
+  r += start_ofs;
+  while (result < tile_end) {
     *result = meta::impl(l, *r);
-    result += blockDim.x * gridDim.x;
-    r += blockDim.x * gridDim.x;
+    result += stride;
+    r += stride;
   }
 }
 
@@ -80,12 +89,15 @@ __device__ void ExecuteBinOp(Result *result, Left l, const Right *r, int64_t ext
 template <ArithmeticOp op, typename Result, typename Left, typename Right>
 __device__ void ExecuteBinOp(Result *result, const Left *l, Right r, int64_t extent) {
   using meta = arithm_meta<op, GPUBackend>;
-  result += blockIdx.x * blockDim.x + threadIdx.x;
-  l += blockIdx.x * blockDim.x + threadIdx.x;
-  for (int64_t i = blockIdx.x * blockDim.x + threadIdx.x; i < extent; i += blockDim.x * gridDim.x) {
+  int64_t start_ofs = static_cast<int64_t>(blockDim.x) * blockIdx.x + threadIdx.x;
+  int64_t stride = static_cast<int64_t>(blockDim.x) * gridDim.x;
+  auto *tile_end = result + extent;
+  result += start_ofs;
+  l += start_ofs;
+  while (result < tile_end) {
     *result = meta::impl(*l, r);
-    result += blockDim.x * gridDim.x;
-    l += blockDim.x * gridDim.x;
+    result += stride;
+    l += stride;
   }
 }