… dali/test/python/test_operator_arithmetic_ops.py