Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>

#### Why we need this PR?
Arithmetic Op is not fast enough, GPU was underutilized due to not big enough tiles.

#### What happened in this PR?
 - What solution was applied:
     Tile was made bigger, grid was made smaller to allow for few more iterations for given thread.
     Pointer usage was adjusted a bit, it helps with small inputs a bit.
     Benchmark was added
 - Affected modules and functionalities:
     Arithmetic Ops
 - Key points relevant for the review:
     
 - Validation and testing:
     Benchmark
 - Documentation (including examples):
     NA


OLD:
```
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'const'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f5fd44f6ae8>, [(1024, 1024)] * 256, '*') ... Throughput: 377.034 GB/s
Throughput: 375.006 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'const'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f5fd44f6ae8>, [(16384, 1024)] * 64, '*') ... Throughput: 391.476 GB/s
Throughput: 405.331 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'const'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f5fd44f6ae8>, [(400, 400)] * 64, '*') ... Throughput: 334.448 GB/s
Throughput: 353.607 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'gpu'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f5fd3e962f0>, [(1024, 1024)] * 256, '*') ... Throughput: 490.778 GB/s
Throughput: 490.198 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'gpu'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f5fd3e962f0>, [(16384, 1024)] * 64, '*') ... Throughput: 499.623 GB/s
Throughput: 507.385 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'gpu'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f5fd3e962f0>, [(400, 400)] * 64, '*') ... Throughput: 358.166 GB/s
Throughput: 402.577 GB/s
ok

----------------------------------------------------------------------
Ran 6 tests in 10.587s

OK
```
NEW:
```
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'const'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f96c80bbae8>, [(1024, 1024)] * 256, '*') ... Throughput: 533.855 GB/s
Throughput: 531.948 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'const'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f96c80bbae8>, [(16384, 1024)] * 64 '*') ... Throughput: 576.293 GB/s
Throughput: 576.419 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'const'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f96c80bbae8>, [(400, 400)] * 64, '*') ... Throughput: 462.535 GB/s
Throughput: 498.008 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'gpu'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f96c7a5c2f0>, [(1024, 1024)] * 256, '*') ... Throughput: 571.668 GB/s
Throughput: 572.734 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'gpu'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f96c7a5c2f0>, [(16384, 1024)] * 64, '*') ... Throughput: 604.232 GB/s
Throughput: 603.718 GB/s
ok
test_operator_arithmetic_ops.test_arithmetic_ops_perf(('gpu', 'gpu'), (<class 'numpy.float32'>, <class 'numpy.float32'>), <function test_arithmetic_ops_perf.<locals>.<lambda> at 0x7f96c7a5c2f0>, [(400, 400)] * 64, '*') ... Throughput: 447.778 GB/s
Throughput: 459.841 GB/s
ok

----------------------------------------------------------------------
Ran 6 tests in 10.426s

OK
```



**JIRA TASK**: *[Use DALI-1514 or NA]*